refactor
Created new C modules that are dedicated for only the RadiationFieldSetup calculations.
