+++ src/py21cmfast/src/RadiationFieldsSetup.h
@@ -1,0 +1,9 @@
+#ifndef _RADIATIONFIELDSSETUP_H
+#define _RADIATIONFIELDSSETUP_H
+
+#include "OutputStructs.h"
+
+int SetupRadiationFields(float redshift, TsBox *previous_spin_temp,
+                         RadiationFieldsSetup *rad_setup);
+
+#endif
